+++ python/sglang/srt/jax/models/qwen.py
@@ -14,6 +14,7 @@ from sglang.srt.jax.layers.linear import LinearBase
 from sglang.srt.jax.layers.logits_processor import LogitsProcessor
 from sglang.srt.jax.layers.quantization.base_config import QuantizationConfig
 from sglang.srt.model_executor.forward_batch_info import ForwardBatch
+from sglang.srt.jax.utils import flatten_pytree_with_paths, get_expected_param_paths, update_state_recursive
 
 
 class QWenMLP(nnx.Module):
@@ -215,60 +216,10 @@ class QWenLMHeadModel(nnx.Module):
         self.lm_head = ParallelLMHead(vocab_size, config.hidden_size, rngs=rngs)
         self.logits_processor = LogitsProcessor(vocab_size)
     
-    def load_pytree_weights(self, pytree):
-        def flatten_pytree_with_paths(pytree, prefix=""):
-            flat_dict = {}
-            if isinstance(pytree, dict):
-                for key, value in pytree.items():
-                    new_prefix = f"{prefix}.{key}" if prefix else key
-                    if isinstance(value, dict):
-                        flat_dict.update(flatten_pytree_with_paths(value, new_prefix))
-                    else:
-                        flat_dict[new_prefix] = value
-            else:
-                flat_dict[prefix] = pytree
-            return flat_dict
-        
-        def get_expected_param_paths(model_state, prefix=""):
-            expected_paths = set()
-            
-            def traverse_state(state, path=""):
-                if hasattr(state, 'value'):
-                    expected_paths.add(path)
-                elif isinstance(state, dict):
-                    for key, value in state.items():
-                        new_path = f"{path}.{key}" if path else key
-                        traverse_state(value, new_path)
-                        
-            traverse_state(model_state, prefix)
-            return expected_paths
-        
-        def update_state_recursive(current_state, flat_weights, path=""):
-            if hasattr(current_state, 'value'):
-                if path in flat_weights:
-                    current_state.value = flat_weights[path]
-                else:
-                    raise ValueError(f"Missing weight for parameter at path '{path}'")
-            elif hasattr(current_state, 'shape'): 
-                if path in flat_weights:
-                    pass
-            elif isinstance(current_state, dict):
-                for key in current_state:
-                    new_path = f"{path}.{key}" if path else key
-                    update_state_recursive(current_state[key], flat_weights, new_path)
-            elif hasattr(current_state, '__iter__') and hasattr(current_state, '__getitem__'):
-                for key in current_state:
-                    new_path = f"{path}.{key}" if path else key
-                    update_state_recursive(current_state[key], flat_weights, new_path)
-            else:
-                raise TypeError(f"Unsupported state type at path '{path}': {type(current_state)}")
-        
+    def load_pytree_weights(self, pytree):        
         flat_weights = flatten_pytree_with_paths(pytree)
-        
         model_state = nnx.state(self)
-        
         expected_paths = get_expected_param_paths(model_state)
-        
         missing_paths = expected_paths - set(flat_weights.keys())
         if missing_paths:
             raise ValueError(f"Missing weights for parameters: {sorted(missing_paths)}")
